feat(ml): integrate Rishabh's ML stack with hardened async adapter
- Transplant Rishabh's contracts.py, scorer.py, groq_client.py,
  priority_engine.py, regex_engine.py, configs/ into services/backend/src/ml/

BLOCKER 1 FIX — ScoringResult schema unified:
  - priority_score: int → float (DB column updated to Float)
  - severity/category: str → StrEnum (Severity/Category from contracts.py)
  - confidence field added to ScoringResult

BLOCKER 2 FIX — ReasoningBreakdown → AIReasoningBreakdown (Rishabh's schema):
  - regex_contribution / local_model_contribution / groq_contribution (weighted)
  - corroboration_bonus, vulnerability_bonus, time_decay_factor
  - extracted_entities, groq_rationale (max 500 chars), matched_rules
  - fallback_used + confidence added for backend monitoring + UI degraded badge
  - ReasoningBreakdown = AIReasoningBreakdown alias for pipeline.py compat

BLOCKER 3 FIX — interface.py rewritten as async adapter:
  - SosPacket → SignedSOSHeader with safe enum coercion (_safe_severity/_safe_category)
  - decrypted_text: str → DecryptedSOSPayload(text=...)
  - Sync scorer runs in asyncio thread executor (non-blocking)
  - 15s hard timeout via asyncio.wait_for() — no event loop stall
  - LLM output sanitized before DB write (rationale cap, entity cap, type enforce)
  - Safe rule-based fallback on timeout or any exception

PIPELINE — pipeline.py:
  - Queries corroborating report count in 500m/2h window before scoring
  - Passes count to adapter (PriorityEngine uses it for cluster boost)
  - Removed redundant inline fallback block (adapter handles all fallbacks)
  - ai_category stored as str(StrEnum), WS payload category as str(StrEnum)

SECURITY — 4 hardening rules at integration boundary:
  - Enum coercion: garbage severity/category → safe default, never crashes
  - Groq timeout: 15s cap prevents async loop stall
  - Rationale cap: 500 chars max (prompt injection / XSS mitigation)
  - Entity/rule list: 20/30 item cap, 100/50 char item cap, str() enforcement

ML TRAINING (separate from backend runtime):
  - ml/ directory added: dataset pipeline, teacher/student KD, TFLite export
  - Multilingual regex test vectors (36 vectors EN/HI/Hinglish) for Arnav

TESTS — 29/29 passing (was 22):
  - 7 new unit tests in test_ml_adapter.py:
    happy path, Groq timeout fallback, severity coercion, category coercion,
    rationale truncation, entity list sanitization, critical fallback

Config:
  - GROQ_MODEL default: llama3-70b-8192 → llama3-8b-8192

+++ services/backend/src/database/models.py
@@ -134,7 +134,7 @@ class SOSReport(Base):
     payload_enc = Column(Text, nullable=False)
     payload_decrypted = Column(Text, nullable=True)
     
-    priority_score = Column(Integer, nullable=True)
+    priority_score = Column(Float, nullable=True)
     ai_category = Column(String(50), nullable=True)
     ai_reasoning = Column(JSON, nullable=True)
